updating the forms structure so that each form in its own separate file
