+++ users/forms/__init__.py
@@ -1,0 +1,3 @@
+from .attendanceForm import AttendanceForm
+from .profileForm import ProfileForm
+from .userPermissionTagForm import UserPermissionTagForm
